Finish out the parser and properly parse planeswalkers and printings.
Printings need to track the individual card number as the same card can be
printed multiple times with different numbers.

Gatherer uses the power/toughness row to store loyalty so planeswalkers were
being assigned a P/T. Now power and toughness are only assigned to cards with
a slash in that row.